scripts/requirements.txt
requests
PyGithub
geojson